backend/requirements.txt
flask
flask-cors
google-generativeai
groq
python-dotenv
anthropic
